responsive bugs fixed, notification bug fixed

+++ backend/routes/userRoutes.js
@@ -2,7 +2,7 @@ const express = require("express");
 const router = express.Router();
 const authMiddleware = require("../middleware/authmiddleware");
 
-const { createProfile, registerUser, updateProfile, markAllNotification,deleteAllNotifications } = require("../controllers/userCtrl");
+const { createProfile, registerUser, updateProfile, markAllNotification,deleteAllNotifications, getCurrentUser} = require("../controllers/userCtrl");
 
 // Register a new user
 // POST /api/users/register
@@ -18,6 +18,9 @@ router.post('/mark-all-notifications', authMiddleware, markAllNotification);
 
 //Delete All Notifications || POST
 router.post('/delete-all-notifications', authMiddleware, deleteAllNotifications);
+
+// Get Current User || GET /api/users/current-user
+router.get('/current-user', authMiddleware, getCurrentUser);
 
 
 module.exports = router;
